[Refactor][BE] 백엔드 API 코드 전체적인 리팩토링

+++ api/routes/payment/index.ts
@@ -1,17 +1,28 @@
-import errorHandler from 'utils/errorHandler';
-import express, { Request, Response } from 'express';
-import { decodeToken, getAccessToken } from 'utils/jwt';
+import express from 'express';
+
 import paymentService from 'services/payment';
+
+import { decodeToken, getAccessToken } from 'utils/jwt';
+import errorHandler from 'utils/error-handler';
+import errorGenerator from 'utils/error-generator';
 
 const router = express.Router();
 
-router.get('/', async (req: Request, res: Response) => {
+/**
+ * path: /payment
+ *
+ * get: 사용자 결제수단 조회
+ * delete: 결제수단 삭제
+ * post: 결제수단 추가
+ */
+
+router.get('/', async (req, res) => {
   try {
     const accessToken = getAccessToken(req.headers.authorization);
     const { uid: userId } = decodeToken(accessToken);
 
     const result = await paymentService.getUserPayment(userId);
-    res.status(200).json({ data: result });
+    res.status(200).json(result);
   } catch (err) {
     console.log(err);
     const { statusCode, errorMessage } = errorHandler(err.code);
@@ -19,16 +30,22 @@ router.get('/', async (req: Request, res: Response) => {
   }
 });
 
-//결제수단 삭제
-router.delete('/', async (req: Request, res: Response) => {
+router.delete('/', async (req, res) => {
   try {
     const accessToken = getAccessToken(req.headers.authorization);
     const { uid: userId } = decodeToken(accessToken);
 
     const { payment } = req.body;
 
-    const result = await paymentService.deleteUserPayment(userId, payment);
-    res.status(200).json({ success: result });
+    if (!payment) {
+      throw errorGenerator({
+        message: 'invalid-body',
+        code: 'req/invalid-body',
+      });
+    }
+
+    await paymentService.deleteUserPayment(userId, payment);
+    res.status(200).json({});
   } catch (err) {
     console.log(err);
     const { statusCode, errorMessage } = errorHandler(err.code);
@@ -36,16 +53,22 @@ router.delete('/', async (req: Request, res: Response) => {
   }
 });
 
-// 결제수단 추가
-router.post('/', async (req: Request, res: Response) => {
+router.post('/', async (req, res) => {
   try {
     const accessToken = getAccessToken(req.headers.authorization);
     const { uid: userId } = decodeToken(accessToken);
 
     const { payment } = req.body;
 
-    const result = await paymentService.createUserPayment(userId, payment);
-    res.status(200).json({ success: result });
+    if (!payment) {
+      throw errorGenerator({
+        message: 'invalid-body',
+        code: 'req/invalid-body',
+      });
+    }
+
+    await paymentService.createUserPayment(userId, payment);
+    res.status(200).json({});
   } catch (err) {
     console.log(err);
     const { statusCode, errorMessage } = errorHandler(err.code);
